added very basic character adding/removal
Added very basic character adding and removal in SQLiteTestActivity.
This screen can be reached by clicking "Home", then "Manage
Characters".

Type some text and click Add and the character name will be added to
the list and immediately displayed. Click Remove and the first
character will be removed. Changes are automatically saved locally.

Because of how it was implented, the "Hello!" button in the main
activity now deletes every string from the list until there is one
left. If the list is empty then it will add "Danny McSizzle".