Added framework for all the tables
Added framework for all the tables. The BasicInfo SQLite helper class
is causing troubles for some reason with its "alignment" attribute.

Also added framework to write to database in all of the
character-related classes, not yet tested. You pass CharacterDataSource
a Character, CDS calls Character.write, which in turn calls write on
each of its subcomponents.